pyproject.toml
[project]
name = "typer-v0-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "typer>=0.21.1",
]
